Extracting the code for loading multilingual content in `TMultilingualContentList`
